backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
requests==2.32.4
python-dotenv==1.0.1